requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.5
customtkinter==5.2.2
darkdetect==0.8.0
distro==1.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.63.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
linkify-it-py==2.0.3
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
packaging==26.0
pillow==12.1.1
platformdirs==4.9.2
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6-WebEngine==6.10.0
PyQt6-WebEngine-Qt6==6.10.2
PyQt6_sip==13.11.0
python-dotenv==1.2.1
requests==2.32.5
rich==14.3.2
rsa==4.9.1
sniffio==1.3.1
SpeechRecognition==3.14.5
tenacity==9.1.4
textual==8.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uc-micro-py==1.0.3
uritemplate==4.2.0
urllib3==2.6.3
websockets==15.0.1
yt-dlp==2026.2.4
